refactoring: Slide와 Card 분리하기
