--- src/components/CardSlider/index.style.ts
@@ -1,11 +1,0 @@
-import styled from '@emotion/styled';
-
-export const SlideWrapper = styled.div`
-  max-width: 400px;
-`;
-
-export const ItemWrapper = styled.div``;
-
-export const LastItemWrapper = styled.div`
-  cursor: pointer;
-`;
