+++ src/components/Slide/index.style.ts
@@ -1,0 +1,5 @@
+import styled from '@emotion/styled';
+
+export const SlideWrapper = styled.div`
+  max-width: 400px;
+`;
