+++ src/components/index.ts
@@ -4,7 +4,7 @@ export { default as CreditCard } from './CreditCard';
 export { default as InputWrapper } from './InputWrapper';
 export { default as Input } from './Input';
 export { default as Keypad } from './Keypad';
-export { default as CardSlider } from './CardSlider';
+export { default as Slide } from './Slide';
 export { default as CompanySelection } from './CompanySelection';
 export { default as GoogleIcon } from './GoogleIcon';
 export { default as DeemBackground } from './DeemBackground';
