+++ src/pages/Payment/components/CardSelection/index.style.ts
@@ -1,0 +1,3 @@
+import styled from '@emotion/styled';
+
+export const CardItemContainer = styled.div``;
